[issue-6] feat: 공통 반환타입 추가
- #6
