+++ kyumall-client/src/main/java/com/kyumall/kyumallclient/member/MemberController.java
@@ -1,7 +1,7 @@
 package com.kyumall.kyumallclient.member;
 
+import com.kyumall.kyumallclient.response.ResponseWrapper;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -19,8 +19,8 @@ public class MemberController {
    * @return
    */
   @PostMapping("/send-verification-mail")
-  public ResponseEntity<Void> sendVerificationMail(@RequestParam String email) {
+  public ResponseWrapper<Void> sendVerificationMail(@RequestParam String email) {
     memberService.sendVerificationEmail(email);
-    return ResponseEntity.ok().build();
+    return ResponseWrapper.ok();
   }
 }
